requirements1.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.6.0
#asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==24.2.0
av==13.0.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.41.0
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
#comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
datasets==2.16.1
#debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1725269181790/work
#decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
decord==0.6.0
deepspeed==0.14.2
defusedxml==0.7.1
dill==0.3.7
docker-pycreds==0.4.0
docopt==0.6.2
docstring_parser==0.16
einops==0.6.1
einops-exts==0.0.4
#exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
#executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
fastapi==0.115.0
fastjsonschema==2.20.0
filelock==3.16.1
flash-attn==2.6.3
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
gradio_client==0.2.9
h11==0.14.0
hf_transfer==0.1.8
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.25.0
idna==3.10
# importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
# ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython==8.12.3
# jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
# jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1726610684920/work
# jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1710257277185/work
jupyterlab_pygments==0.3.0
latex2mathml==3.77.0
# -e git+https://github.com/TigranGalstyan/LLaVA-NeXT-AirVLN.git@00e3cea3ca5a85b7ecaf6ee01358aa207d170e83#egg=llava
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
# matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
# nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.3
ninja==1.11.1.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.26.1
opencv-python==4.10.0.84
# packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.2.3
pandocfilters==1.5.1
# parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
peft==0.4.0
# pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
# pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
pipreqs==0.5.0
# platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1726613481435/work
# prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
protobuf==5.28.2
# psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1725737838492/work
# ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
# pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pydantic==1.10.8
# Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pynvml==11.5.3
# python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz==2024.2
PyYAML==6.0.2
# pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1725448950123/work
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.2.2
scipy==1.14.1
sentencepiece==0.1.99
sentry-sdk==2.14.0
setproctitle==1.3.3
shortuuid==1.0.13
shtab==1.7.1
# six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
# stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.38.6
svgwrite==1.4.3
sympy==1.13.3
threadpoolctl==3.5.0
timm==1.0.9
tinycss2==1.3.0
tokenizers==0.15.2
torch==2.1.2
torchvision==0.16.2
# tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1724960687225/work
tqdm==4.66.5
# traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
# transformers @ git+https://github.com/huggingface/transformers.git@1c39974a4c4036fd641bc1191cc32799f85715a4
triton==2.1.0
# typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tyro==0.8.11
tzdata==2024.1
urllib3==1.26.20
uvicorn==0.30.6
wandb==0.18.1
wavedrom==2.0.3.post3
# wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
webencodings==0.5.1
websockets==13.1
xxhash==3.5.0
yarg==0.1.9
yarl==1.11.1
# zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1726248574750/work
